ship moved to packages, Position renamed
