+++ src/main/java/at/ac/fhcampuswien/battleship/Position.java
@@ -1,0 +1,23 @@
+package at.ac.fhcampuswien.battleship;
+
+public class AttackPositions {
+
+    private int x;
+
+    private int y;
+
+    public AttackPositions(int x, int y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public int getX()
+    {
+        return x;
+    }
+
+    public int getY()
+    {
+        return y;
+    }
+}
